Errors fix

+++ ApiVuelos/src/main/java/com/ApiVuelos/ApiVuelos/controller/CountryController.java
@@ -17,88 +17,112 @@ public class CountryController {
     @Autowired
     private CountryService countryService;
 
-    @PostMapping(value = "/" ,consumes="application/json")
-    public ResponseEntity add(@RequestParam("name")String name,@RequestParam("iso")String isoCode) {
+    @PostMapping(value = "/" , consumes = "application/json")
+    public ResponseEntity add(@RequestBody List<Country> countries) {
+
+        ResponseEntity status = new ResponseEntity(HttpStatus.INTERNAL_SERVER_ERROR);
+
         try{
-            if(name!=null && isoCode!=null){
-                Country country = new Country(name, isoCode);
-                this.countryService.newObject(country);
-                return new ResponseEntity(HttpStatus.OK);
-            }
-            else
-            {
-                return new ResponseEntity(HttpStatus.NO_CONTENT);
+            for(Country country : countries) {
+                if (country.getName() != null && !(country.getName().trim().equals("")) && country.getIsoCode() != null && !(country.getIsoCode().trim().equals(""))) {
+                    this.countryService.newObject(country);
+                    status = new ResponseEntity(HttpStatus.OK);
+
+                } else {
+                    status = new ResponseEntity(HttpStatus.NO_CONTENT);
+                }
             }
         }
         catch(Exception e){
-            return new ResponseEntity(HttpStatus.INTERNAL_SERVER_ERROR);
+            status = new ResponseEntity(HttpStatus.INTERNAL_SERVER_ERROR);
         }
+
+        return status;
     }
 
     @PutMapping(value = "/")
     public ResponseEntity update(Country country){
+
+        ResponseEntity status = new ResponseEntity(HttpStatus.INTERNAL_SERVER_ERROR);
+
         try{
             if(country!=null){
                 this.countryService.updateObject(country);
-                return new ResponseEntity(HttpStatus.OK);
+                status = new ResponseEntity(HttpStatus.OK);
             }
             else{
-                return new ResponseEntity(HttpStatus.NO_CONTENT);
+                status = new ResponseEntity(HttpStatus.NO_CONTENT);
             }
         }
         catch(Exception e){
-            return new ResponseEntity(HttpStatus.INTERNAL_SERVER_ERROR);
+            status = new ResponseEntity(HttpStatus.INTERNAL_SERVER_ERROR);
         }
+        return status;
     }
 
     @DeleteMapping(value = "/")
     public ResponseEntity remove(@RequestParam("id") Long id){
+
+        ResponseEntity status = new ResponseEntity(HttpStatus.INTERNAL_SERVER_ERROR);
+
         try{
-            if(id!=null){
+            if(id != null && id != 0){
                 this.countryService.removeObject(id);
-                return new ResponseEntity(HttpStatus.OK);
+                status = new ResponseEntity(HttpStatus.OK);
             }else{
-                return new ResponseEntity(HttpStatus.NO_CONTENT);
+                status = new ResponseEntity(HttpStatus.NO_CONTENT);
             }
         }
         catch(Exception e){
-            return new ResponseEntity(HttpStatus.INTERNAL_SERVER_ERROR);
+            status = new ResponseEntity(HttpStatus.INTERNAL_SERVER_ERROR);
         }
+        return status;
     }
 
     @GetMapping
     public ResponseEntity<List<Country>> getAll() {
-        List<Country> countryList=new ArrayList<Country>();
+
+        ResponseEntity status = new ResponseEntity(HttpStatus.INTERNAL_SERVER_ERROR);
+        List<Country> countryList = new ArrayList<Country>();
+
         try{
            countryList = this.countryService.getAll();
-           if(countryList!=null){
-               return new ResponseEntity<List<Country>>(countryList,HttpStatus.OK);
-           }else
-           {
-               return new ResponseEntity<List<Country>>(HttpStatus.NO_CONTENT);
+           if(!countryList.isEmpty()){
+               status = new ResponseEntity<List<Country>>(countryList,HttpStatus.OK);
+
+           }else {
+               status = new ResponseEntity<List<Country>>(HttpStatus.NO_CONTENT);
            }
         }
         catch(Exception e){
-            return new ResponseEntity<List<Country>>(HttpStatus.INTERNAL_SERVER_ERROR);
+            status = new ResponseEntity<List<Country>>(HttpStatus.INTERNAL_SERVER_ERROR);
         }
+        return status;
     }
+
     @GetMapping(value="/")
     public ResponseEntity getByOneCountry(@RequestParam("iso") String iso){
-        Country count=null;
+
+        ResponseEntity status = new ResponseEntity(HttpStatus.INTERNAL_SERVER_ERROR);
+        Country country = null;
+
         try{
             if(iso!=null){
-                count= this.countryService.getByAttributeType(iso);
-                if(count!=null){
-                    return new ResponseEntity(count,HttpStatus.OK);
+                country= this.countryService.getByAttributeType(iso);
+                if(country!=null){
+                    status = new ResponseEntity(country,HttpStatus.OK);
+
+                }   else {
+                    status = new ResponseEntity(HttpStatus.NO_CONTENT);
                 }
             }else{
-                return new ResponseEntity(HttpStatus.NO_CONTENT);
+                status = new ResponseEntity(HttpStatus.NO_CONTENT);
             }
         }
         catch(Exception e)
         {
-           return new ResponseEntity(HttpStatus.INTERNAL_SERVER_ERROR);
+           status = new ResponseEntity(HttpStatus.INTERNAL_SERVER_ERROR);
         }
-        return new ResponseEntity(HttpStatus.NO_CONTENT);
+        return status;
     }
 }
